packages/train/pyproject.toml
[project]
name = "curation-train"
version = "0.3.0"
description = "Qwen2.5-VL multimodal finetune runner for the curation pipeline"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.2",
    "torchvision",
    "transformers",
    "deepspeed>=0.12.6",
    "flash-attn>=2.0,<2.8",
    "datasets",
    "pillow",
    "accelerate",
    "peft",
    "huggingface_hub",
    "qwen-vl-utils",
    "packaging",
    "setuptools>=82.0.0",
]

[project.scripts]
curation-train = "curation_train.cli:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[tool.uv.extra-build-dependencies]
flash-attn = ["torch"]

[tool.uv]
package = true
no-binary-package = ["flash-attn"]
